Allow points in halos for _pic_swarm_in_mesh
Currently switched off when creating a VOM but tested in test_swarm.py.
I'll need to update the SF for the VOM to allow this to be switched on.

It's useful to add this now since I'll be fiddling with the VOM
embedding code, so the tests will tell me if I break anything.