Save the rank that a point is on as "onrank" field
I've added this so that the _vom_original_ordering function will permute
this information too, in case it's needed. If this is overkill, I can
revert this.

I wonder if this is what the DMSwarm_rank field should have saved in it,
and this should be renamed "owned_rank" or similar.