Improve remote index search algorithm
I've switched to finding the intersection of two arrays, which I do once
for each rank in the halo (rather than for each particle on each rank in
the halo!).

I've also made the search space a bit smaller: we only need to search
for matches between halos and owned particles.